Reorganize test scripts according to the structure described in the README file.
