+++ test/commands/camera/test_camera.py
