+++ test/ipc/test_connection.py
